+++ src/components/Form/FormSelectSingle.tsx
@@ -22,7 +22,7 @@ export const FormSelectSingle = <
     options,
     selectedItems,
     setSelectedItems,
-    valueMode: formValueMode ?? 'id',
+    valueMode: formValueMode,
   });
   const selectedItem = selectedItems[0] ?? null;
   return (