+++ src/components/Form/useMultiSelectFormSync.ts
@@ -1,3 +1,4 @@
+import { useState } from 'react';
 import { FieldValues, useFormContext } from 'react-hook-form';
 import { GenericDataType } from '../../common';
 import { useValueChangeEffect } from '../../hooks';
@@ -15,34 +16,38 @@ export const useMultiSelectFormSync = <
   valueMode,
 }: SelectSyncOptions<Values, DataItem>) => {
   const { watch, setValue } = useFormContext();
+  const [isDefaultLoaded, setIsDefaultLoaded] = useState(false);
   const values = watch(name);
   useValueChangeEffect(
     values,
     () => {
-      if (valueMode === 'item') {
-        setSelectedItems((values as DataItem[]) ? values : []);
-      } else {
+      if (valueMode === 'id') {
         const ids = values as DataItem['id'][];
         setSelectedItems(
           ids ? ids.flatMap(id => (options[id] ? options[id][0] : [])) : [],
         );
+      } else {
+        setSelectedItems((values as DataItem[]) ? values : []);
       }
+      setIsDefaultLoaded(true);
     },
     [valueMode, options],
   );
   useValueChangeEffect(
     selectedItems,
     () => {
-      const itemValues =
-        valueMode === 'item'
-          ? selectedItems
-          : selectedItems.map(({ id }) => id);
-      setValue<string>(name, itemValues, {
-        shouldTouch: true,
-        shouldDirty: true,
-      });
-      onItemSelect?.(selectedItems);
+      if (isDefaultLoaded) {
+        const itemValues =
+          valueMode === 'id'
+            ? selectedItems.map(({ id }) => id)
+            : selectedItems;
+        setValue<string>(name, itemValues, {
+          shouldTouch: true,
+          shouldDirty: true,
+        });
+        onItemSelect?.(selectedItems);
+      }
     },
-    [onItemSelect, valueMode],
+    [isDefaultLoaded, onItemSelect, valueMode],
   );
 };