+++ src/components/Select/Select.stories.tsx
@@ -56,7 +56,14 @@ Default.args = {
 export const WithDefaultValue: Story<
   SelectProps<DataItem, FormValues>
 > = args => {
-  const methods = useForm();
+  const methods = useForm({
+    defaultValues: {
+      field1: {
+        id: '2',
+        label: 'Item 2',
+      },
+    },
+  });
   return (
     <Form className="h-60 w-60" methods={methods} onFormSubmit={() => null}>
       <Select {...args} />
@@ -96,6 +103,51 @@ export const SelectMultiple: Story<
 };
 
 SelectMultiple.args = {
+  labelText: 'Field Label',
+  multi: true,
+  name: 'field1',
+  getItemText: ({ id, label }) => `${id} - ${label}`,
+  options: [
+    {
+      id: '1',
+      label: 'Item 1',
+    },
+    {
+      id: '2',
+      label: 'Item 2',
+    },
+    {
+      id: '3',
+      label: 'Item 3',
+    },
+  ],
+};
+
+export const WithMultipleDefaultValues: Story<
+  SelectProps<DataItem, FormValues>
+> = args => {
+  const methods = useForm({
+    defaultValues: {
+      field1: [
+        {
+          id: '2',
+          label: 'Item 2',
+        },
+        {
+          id: '3',
+          label: 'Item 3',
+        },
+      ],
+    },
+  });
+  return (
+    <Form className="h-60 w-60" methods={methods} onFormSubmit={() => null}>
+      <Select {...args} />
+    </Form>
+  );
+};
+
+WithMultipleDefaultValues.args = {
   labelText: 'Field Label',
   multi: true,
   name: 'field1',